+++ src/integration-test/java/org/openlmis/referencedata/SqlInsert.java
@@ -42,8 +42,8 @@ final class SqlInsert {
   };
 
   private static final String[] FTAP_FIELDS = {
-      "id", "emergencyorderpoint", "maxperiodsofstock", "minperiodsofstock", "facilitytypeid",
-      "orderableid", PROGRAM_ID
+      "id", "versionId", "emergencyorderpoint", "maxperiodsofstock", "minperiodsofstock",
+      "facilitytypeid", "orderableid", PROGRAM_ID
   };
 
   private static final String[] FACILITY_TYPE_FIELDS = {